+ Binds are now saved to slade.cfg + Added input configuration to the preferences dialog + Added 'splash' command to the console, initially for testing, but added some fun stuff :P * Hopefully fixed 3d mode flickering, again :P + Added a proper splash screen (instead of just popping up the console like before) ' Moved over to VC++ 2005 Express, now intellisense works with GTK. No more constantly checking the API docs ;)

+++ trunk/draw.cpp
@@ -495,8 +495,8 @@ void draw_sectors()
 
 	for(DWORD l = 0; l < map.n_lines; l++)
 	{
-		short sector1 = map.l_getsector1(l);
-		short sector2 = map.l_getsector2(l);
+		long sector1 = map.l_getsector1(l);
+		long sector2 = map.l_getsector2(l);
 
 		rect_t rect = map.l_getsrect(l);
